Remove class syntax from controllers
