+++ backend/api/routes/configs.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 import { authenticateRequest } from "../../middlewares/auth";
 import { asyncHandler, validateRequest } from "../../middlewares/api-utils";
 import configSchema from "../schemas/config-schema";
-import ConfigController from "../controllers/config";
+import * as ConfigController from "../controllers/config";
 import * as RateLimit from "../../middlewares/rate-limit";
 
 const router = Router();
