+++ backend/api/routes/presets.ts
@@ -1,6 +1,6 @@
 import joi from "joi";
 import { authenticateRequest } from "../../middlewares/auth";
-import PresetController from "../controllers/preset";
+import * as PresetController from "../controllers/preset";
 import * as RateLimit from "../../middlewares/rate-limit";
 import configSchema from "../schemas/config-schema";
 import { asyncHandler, validateRequest } from "../../middlewares/api-utils";
