+++ backend/api/routes/quotes.ts
@@ -1,7 +1,7 @@
 import joi from "joi";
 import { authenticateRequest } from "../../middlewares/auth";
 import { Router } from "express";
-import QuotesController from "../controllers/quotes";
+import * as QuoteController from "../controllers/quote";
 import * as RateLimit from "../../middlewares/rate-limit";
 import {
   asyncHandler,
@@ -23,7 +23,7 @@ quotesRouter.get(
   RateLimit.newQuotesGet,
   authenticateRequest(),
   checkIfUserIsQuoteMod,
-  asyncHandler(QuotesController.getQuotes)
+  asyncHandler(QuoteController.getQuotes)
 );
 
 quotesRouter.post(
@@ -46,7 +46,7 @@ quotesRouter.post(
     },
     validationErrorMessage: "Please fill all the fields",
   }),
-  asyncHandler(QuotesController.addQuote)
+  asyncHandler(QuoteController.addQuote)
 );
 
 quotesRouter.post(
@@ -62,7 +62,7 @@ quotesRouter.post(
     validationErrorMessage: "Please fill all the fields",
   }),
   checkIfUserIsQuoteMod,
-  asyncHandler(QuotesController.approveQuote)
+  asyncHandler(QuoteController.approveQuote)
 );
 
 quotesRouter.post(
@@ -75,7 +75,7 @@ quotesRouter.post(
     },
   }),
   checkIfUserIsQuoteMod,
-  asyncHandler(QuotesController.refuseQuote)
+  asyncHandler(QuoteController.refuseQuote)
 );
 
 quotesRouter.get(
@@ -88,7 +88,7 @@ quotesRouter.get(
       language: joi.string().required(),
     },
   }),
-  asyncHandler(QuotesController.getRating)
+  asyncHandler(QuoteController.getRating)
 );
 
 quotesRouter.post(
@@ -102,7 +102,7 @@ quotesRouter.post(
       language: joi.string().required(),
     },
   }),
-  asyncHandler(QuotesController.submitRating)
+  asyncHandler(QuoteController.submitRating)
 );
 
 quotesRouter.post(
@@ -137,7 +137,7 @@ quotesRouter.post(
       return !user.cannotReport;
     },
   }),
-  asyncHandler(QuotesController.reportQuote)
+  asyncHandler(QuoteController.reportQuote)
 );
 
 export default quotesRouter;
